# app/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: QuickEx Backend API
  description: >
    QuickEx is a Stellar-based payment platform enabling payment links,
    recurring subscriptions, username resolution, marketplace trading,
    fiat on/off ramps, and webhook-driven event notifications.

    **Base URL (production):** `https://api.quickex.example.com`
    **Base URL (development):** `http://localhost:3000`

    Interactive Swagger UI is available at `/docs` when the server is running.
  version: 1.0.0
  contact:
    name: QuickEx API Support
    url: https://github.com/quickex/quickex
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.quickex.example.com
    description: Production
...
security:
  - ApiKeyAuth: []
  - {}
paths:
...
  /username:
    post:
      tags: [usernames]
      operationId: createUsername
      summary: Create a new username
      description: >
        Registers a new username for a Stellar wallet. Username must be 3-32 characters,
        lowercase alphanumeric with underscores. Duplicate usernames return 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUsernameRequest"
      responses:
        "201":
          description: Username created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUsernameResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Wallet has reached maximum usernames
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [usernames]
      operationId: listUsernames
      summary: List usernames for a wallet
      description: Returns all usernames registered for the given Stellar public key.
      parameters:
        - name: publicKey
          in: query
          required: true
          schema:
            type: string
          description: Stellar public key (G...)
      responses:
        "200":
          description: List of usernames
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListUsernamesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /username/search:
    get:
      tags: [usernames]
      operationId: searchUsernames
      summary: Search public profiles (fuzzy)
      description: >
        Fuzzy search for public usernames sorted by similarity score.
        Only profiles with public visibility enabled appear in results.
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Search term (min 2 characters)
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: cursor
          in: query
          schema:
            type: string
          description: Opaque pagination cursor
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchUsernamesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /username/trending:
    get:
      tags: [usernames]
      operationId: getTrendingCreators
      summary: Get trending creators
      description: Returns trending creator profiles ranked by transaction volume.
      parameters:
        - name: timeWindowHours
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 720
            default: 24
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Trending creators
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendingCreatorsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /username/recently-active:
    get:
      tags: [usernames]
      operationId: getRecentlyActive
      summary: Get recently active users
      description: Returns users recently active based on payment activity.
      parameters:
        - name: timeWindowHours
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 168
            default: 24
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Recently active users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentlyActiveResponse"
  /username/toggle-public:
    post:
      tags: [usernames]
      operationId: togglePublicProfile
      summary: Toggle public profile visibility
      description: Enable or disable public profile visibility for a username.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TogglePublicRequest"
      responses:
        "200":
          description: Visibility toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Links
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for higher rate limits

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
...
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            request_id:
              type: string
              format: uuid
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            details:
              type: object

    # --- Health ---
...
    CreateUsernameRequest:
      type: object
      required: [username, publicKey]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[a-z0-9_]+$"
          example: alice_123
        publicKey:
          type: string
          example: GBXGQ55JMQ4L2B6E7S8Y9Z0A1B2C3D4E5F6G7H8I7YWR
    CreateUsernameResponse:
      type: object
      properties:
        ok:
          type: boolean
    ListUsernamesResponse:
      type: object
      properties:
        usernames:
          type: array
          items:
            type: object
            properties:
              username:
                type: string
              publicKey:
                type: string
              isPublic:
                type: boolean
    PublicProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        publicKey:
          type: string
        similarityScore:
          type: number
        lastActiveAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    SearchUsernamesResponse:
      type: object
      properties:
        profiles:
          type: array
          items:
            $ref: "#/components/schemas/PublicProfile"
        total:
          type: integer
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
    TrendingCreatorsResponse:
      type: object
      properties:
        creators:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/PublicProfile"
              - type: object
                properties:
                  transactionVolume:
                    type: number
                  transactionCount:
                    type: integer
        timeWindowHours:
          type: integer
        calculatedAt:
          type: string
          format: date-time
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
    RecentlyActiveResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/PublicProfile"
        timeWindowHours:
          type: integer
        calculatedAt:
          type: string
          format: date-time
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
    TogglePublicRequest:
      type: object
      required: [username, publicKey, isPublic]
      properties:
        username:
          type: string
          example: alice
        publicKey:
          type: string
        isPublic:
          type: boolean

    # --- Links ---
...